Signal Spy: prefer size_t instead of unsigned int
It removes a bunch of warnings in the usual case where you use the size of
the records to know how many records you are expecting next.

Reviewed-on: http://gerrit.aldebaran.lan/70695
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>

+++ qi/signalspy.hpp
@@ -117,7 +117,7 @@ public:
   }
 
   /// Waits for the given number of records to be reached, before the given timeout.
-  qi::FutureSync<bool> waitUntil(unsigned int nofRecords, const qi::Duration& timeout) const
+  qi::FutureSync<bool> waitUntil(size_t nofRecords, const qi::Duration& timeout) const
   {
     qi::Promise<bool> waiting;
     async([this, waiting, nofRecords, timeout]() mutable
